Crate wax

Wax provides opinionated and portable globs that can be matched against file paths and directory trees. Globs use a familiar syntax and support expressive features with semantics that emphasize component boundaries.

See the repository documentation for details about glob expressions and patterns.

Modules

prelude

Re-exports of commonly used items.

walkwalk

Traversal and matching of files and directory trees.

Structs

Any

Combinator that matches any of its component Programs.

BuildError

Describes errors that occur when building a Program from a glob expression.

CandidatePath

Path that can be matched against a Program.

CapturingToken

Token that captures matched text in a glob expression.

Glob

Program that can be matched against paths and directory trees.

MatchedText

Text that has been matched by a Program and its captures.

Enums

GlobError

General errors concerning Programs.

Variance

Variance of a Program.

Traits

LocatedError

Error associated with a Span within a glob expression.

Pattern

A representation of a glob expression.

Program

A compiled glob expression that can be inspected and matched against paths.

Functions

any

Constructs a combinator that matches if any of its input Patterns match.

escape

Escapes text as a literal glob expression.

is_contextual_meta_character

Returns true if the given character is a contextual meta-character.

is_meta_character

Returns true if the given character is a meta-character.

Type Aliases

Span

Location and length of a token within a glob expression.