Module wax::prelude

Re-exports of commonly used items.

This module anonymously re-exports traits for matching Programs against file paths and directory trees. A glob import of this module can be used instead of individual imports of these traits.

Examples

use wax::prelude::*;
use wax::Glob;

// This code requires the `Entry` and `FileIterator` traits.
let glob = Glob::new("**/*.(?i){jpg,jpeg}").unwrap();
for entry in glob.walk("textures").not(["**/.*/**"]).unwrap().flatten() {
    println!("JPEG: {:?}", entry.path());
}

Re-exports

pub use crate::walk::Entry as _;

pub use crate::walk::FileIterator as _;

pub use crate::walk::PathExt as _;

pub use crate::LocatedError as _;

pub use crate::Program as _;