Trait wax::LocatedError

pub trait LocatedError: Display {
    // Required method
    fn span(&self) -> Span;
}

Error associated with a Span within a glob expression.

Located errors describe specific instances of an error within a glob expression. Types that implement this trait provide a location within a glob expression via the LocatedError::span function as well as a description via the Display trait. See BuildError::locations.

Required Methods

fn span(&self) -> Span

Gets the span within the glob expression with which the error is associated.