Struct wax::walk::WalkBehavior

pub struct WalkBehavior {
    pub depth: usize,
    pub link: LinkBehavior,
}

Available on crate feature walk only.

Configuration for walking directory trees.

Determines the behavior of the traversal within a directory tree when using functions like Glob::walk_with_behavior. WalkBehavior can be constructed via conversions from types representing its fields. APIs generally accept impl Into<WalkBehavior>, so these conversion can be used implicitly. When constructed using such a conversion, WalkBehavior will use defaults for any remaining fields.

Examples

By default, symbolic links are interpreted as regular files and targets are ignored. To read linked targets, use LinkBehavior::ReadTarget.

use wax::walk::LinkBehavior;
use wax::Glob;

for entry in Glob::new("**")
    .unwrap()
    .walk_with_behavior(".", LinkBehavior::ReadTarget)
{
    let entry = entry.unwrap();
    // ...
}

Fields
depth: usize

Maximum depth.

Determines the maximum depth to which a directory tree will be traversed relative to the root. A depth of zero corresponds to the root and so using such a depth will yield at most one file entry that refers to the root.

For PathExt::walk, this depth is relative to the Path receiver. For Glob::walk, this depth is relative to the directory path parameter.

The default value is usize::MAX.

link: LinkBehavior

Interpretation of symbolic links.

Determines how symbolic links are interpreted when walking a directory tree. See LinkBehavior.

The default value is LinkBehavior::ReadFile.